Change spline behaviour at edges
In the old function, every point is represented by three splines except the points at the edges, which are represented by only two splines. 
This function changes the construction of the spline basis such that every point is represented by three splines. This comes at the cost that the spline kernels have wider support, giving a coarser approximation.